Changed LeaderCard and BaseCard into one component, LeaderBaseCard.
Deleted the BaseCardProps and LeaderCardProps types from the global file, and defined LeaderBaseCardProps.

I also made sure the yarn.lock was correct as I had accidentally updated socket.io-client to 4.8.1 when I yarn installed, so I put it back to what you had before, 4.8.0. I had already caught it but only modified the package.json in the previous commit.
